Factor TempData and HumidityData classes into separate modules
